chore: move the provider creation to its own module
