+++ engine/crates/tracing/src/otel.rs
@@ -1,5 +1,7 @@
 /// Contains opentelemetry tracing integrations, namely [tracing_subscriber::Layer]'s and
 pub mod layer;
+/// For creation of a tracing provider.
+pub mod provider;
 
 // re-exporting otel libs
 pub use opentelemetry;
